Revert changes
