+++ packages/typespec-ts/test/unit/modelsGenerator.spec.ts
@@ -4,7 +4,7 @@ import {
   emitParameterFromTypeSpec,
   emitResponsesFromTypeSpec
 } from "../util/emitUtil.js";
-import { assertEqualContent, verifyPropertyType } from "../util/testUtil.js";
+import { VerifyPropertyConfig, assertEqualContent } from "../util/testUtil.js";
 
 describe("Input/output model type", () => {
   it("shouldn't generate models if there is no operations", async () => {
@@ -19,6 +19,70 @@ describe("Input/output model type", () => {
     assert.ok(!inputModelFile);
     assert.ok(!outputModelFile);
   });
+
+  async function verifyPropertyType(
+    tspType: string,
+    inputType: string,
+    options?: VerifyPropertyConfig,
+    needAzureCore: boolean = false,
+    additionalImports: string = ""
+  ) {
+    const defaultOption: VerifyPropertyConfig = {
+      additionalTypeSpecDefinition: "",
+      outputType: inputType,
+      additionalInputContent: "",
+      additionalOutputContent: ""
+    };
+    const {
+      additionalTypeSpecDefinition,
+      outputType,
+      additionalInputContent,
+      additionalOutputContent
+    } = {
+      ...defaultOption,
+      ...options
+    };
+    const schemaOutput = await emitModelsFromTypeSpec(
+      `
+    ${additionalTypeSpecDefinition}
+    #suppress "@azure-tools/typespec-azure-core/documentation-required" "for test"
+    model InputOutputModel {
+      prop: ${tspType};
+    }
+  
+    #suppress "@azure-tools/typespec-azure-core/use-standard-operations" "for test"
+    #suppress "@azure-tools/typespec-azure-core/documentation-required" "for test"
+    @route("/models")
+    @get
+    op getModel(@body input: InputOutputModel): InputOutputModel;`,
+      needAzureCore
+    );
+    assert.ok(schemaOutput);
+    const { inputModelFile, outputModelFile } = schemaOutput!;
+    assert.strictEqual(inputModelFile?.path, "models.ts");
+    assertEqualContent(
+      inputModelFile?.content!,
+      `
+    ${additionalImports}
+  
+    export interface InputOutputModel {
+        prop: ${inputType};
+    }
+    ${additionalInputContent}`
+    );
+
+    assert.strictEqual(outputModelFile?.path, "outputModels.ts");
+    assertEqualContent(
+      outputModelFile?.content!,
+      `
+    ${additionalImports}
+  
+    export interface InputOutputModelOutput {
+      prop: ${outputType};
+    }
+    ${additionalOutputContent}`
+    );
+  }
 
   describe("null generation", async () => {
     it("should generate null only", async () => {
